refactor: table cell component
