+++ main/npm/dependency/getInfo.ts
@@ -1,18 +1,13 @@
+import packageJSON, { AbbreviatedMetadata } from 'package-json';
 import executeCommandJSON from '../../utils/executeCommandJSON';
 import log from '../../utils/log';
 import nodeCache from '../../utils/nodeCache';
+import { getCurrentRegistry } from '../registry';
 
 interface InstalledDependency {
   version: string;
   from?: string;
   resolved?: string;
-}
-
-interface OutdatedDependency {
-  current: string;
-  wanted: string;
-  latest: string;
-  location: string;
 }
 
 interface InstalledDependencies {
@@ -29,18 +24,24 @@ export async function getGlobalDependencies(force: boolean) {
       return cache;
     }
     const { dependencies: installedDeps } = await executeCommandJSON<InstalledDependencies>('npm', ['list', '-g', '--depth=0', '--json']);
-    const outdatedDeps = await executeCommandJSON('npm', ['outdated', '-g', '--json']);
+    // exclude the dependency which the version is null and the ignore dependencies
+    const deps = Object.keys(installedDeps)
+      .filter((name: string) => !IGNORE_DEPENDENCIES.includes(name) && installedDeps[name].version);
 
-    const depsInfo = Object.keys(installedDeps)
-      .filter((name: string) => !IGNORE_DEPENDENCIES.includes(name) && installedDeps[name].version)
-      .map((name: string) => {
-        return {
-          name,
-          type: 'global',
-          currentVersion: getCurrentVersion(installedDeps[name]),
-          latestVersion: getLatestVersion(outdatedDeps[name]),
-        };
+    const depsInfo = [];
+    for (const dep of deps) {
+      const latestVersion = await getLatestVersion(dep);
+      const currentVersion = getCurrentVersion(installedDeps[dep]);
+
+      depsInfo.push({
+        name: dep,
+        type: 'global',
+        currentVersion,
+        latestVersion,
+        isOutdated: latestVersion !== currentVersion,
       });
+    }
+
     log.info('depsInfo: ', depsInfo);
     nodeCache.set(GLOBAL_DEPS_KEY, depsInfo);
     return depsInfo;
@@ -53,7 +54,16 @@ export async function getGlobalDependencies(force: boolean) {
 function getCurrentVersion(installedDependency: InstalledDependency) {
   return installedDependency.version;
 }
+interface PackageJSON extends AbbreviatedMetadata {
+  version: string;
+}
 
-function getLatestVersion(outdatedDependency?: OutdatedDependency) {
-  return !outdatedDependency ? '' : outdatedDependency.latest;
+async function getLatestVersion(name: string) {
+  const registryUrl = await getCurrentRegistry();
+  const { version: latest } = await packageJSON(
+    name,
+    { registryUrl },
+  ) as PackageJSON;
+
+  return latest;
 }
