+++ renderer/src/pages/Node/components/LatestVersionTabelCell/index.module.scss
@@ -1,0 +1,14 @@
+.columnCell {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+
+  .button {
+    margin-left: 10px;
+    position: absolute;
+    left: 70px;
+    top: -2px;
+  }
+}
