+++ renderer/src/pages/Node/components/NpmDependency/index.module.scss
@@ -11,16 +11,6 @@
 
 .table {
   margin-top: 20px;
-
-  .columnCell {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-}
-
-.button {
-  margin-left: 10px;
 }
 
 .placeholder {
