+++ renderer/src/pages/Node/components/OperationTableCell/index.module.scss
@@ -1,0 +1,9 @@
+.columnCell {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  .button {
+    margin-left: 10px;
+  }
+}
